[Res] fix | fix errors (#36)
* [PROCJET] fix | remove std::cout print location address, remove unused value in FileSystem, change for (size_t k = 0; k < locations.size(); j++)  -> k++
resolved include error and virtual error

* [RES] feat | impliment response Hnadle

* feat: Remove unused files and refactor directory structure

* [RES] feat | impliment ResponseHandle class

* [RES] refactor | refactor ResponseHandle change filapathFunction, change Constructor
Response handlers are created with port as an argument

* [RES] refactor | Remove unnecessary code in ExistDir.cpp and getServer function

* [RES] Update | common/FileSystem/GetFileSize.cpp

Co-authored-by: Yechan kim <dpcks5959@naver.com>

* [RES] feat | Add ResponseUtils.hpp and setReasonPhrase() function

* [RES] refactor | Remove Exist function and update file paths

* [RES] refactor | Add response action, add setcookie function
Manage getExtesion() function as a map

* [RES] refactor | Remove unused code in Client.cpp and ResponseHandle.cpp

* [REQ] refactor | Trim key and value in Request::parseRequestHeader

The code changes refactor the `Request::parseRequestHeader` function to correctly trim the `key` and `value` variables before further processing. This ensures that any leading or trailing whitespace is removed, preventing potential issues with invalid request header formats.

* [RES] refactor | Remove unused variables in responseHandle.hpp

* [RES ]refactor | Remove unused getUserId() function in Client.hpp

* [CLIENT] feat | Add isKeepAlive() function to Client class

The code changes add the `isKeepAlive()` function to the `Client` class. This function checks if the request header contains the "Connection" field with a value of "keep-alive". This is useful for determining if the client wants to keep the connection open for multiple requests.

* [RES] feat | Update Cache-Control header to use "public, max-age=3600"

* [RES] refactor | Remove unused code and update file paths in responseHandle.hpp and Client.cpp

* [RES] feat | Remove setCookie function and unused code in response.hpp and response.cpp

* [RES] refactor | Remove unused code and update response assignment operator

* [RES] refactor | Remove unused code and update response assignment operator

* [RES] refactor | Update file paths and remove unused code in responseHandle.hpp and Client.cpp

* [RES] feat | Add new functions to FileSystem namespace

The code changes add new functions to the `FileSystem` namespace, including `ToString` for converting a path to a string, `getFolderInfo` for retrieving information about a folder, and `canAccessFolder` for checking if a user has access to a folder.

---------

Co-authored-by: Yechan kim <dpcks5959@naver.com>