[SERVER] fix | fix server event error & remove Debug print (#43)
* [RES/SOCKET] HOTFIX | resolved the socket problome

* fix: change allow_methods to POST in default.conf

The allow_methods directive in the default.conf file was changed from GET to POST. This allows the server to accept POST requests in addition to GET requests.

* remove exe file

* [SERVER] feat: Remove debug print statements and update server socket backlog

The debug print statements in the `Client.cpp` and `ResponseHandle.cpp` files were commented out to remove unnecessary console output. Additionally, the server socket backlog in the `ServerSocket.cpp` file was updated to use the `SOMAXCONN` constant, which allows for a larger number of pending connections.