[SERVER] feat | ignore signal SIGINT SIGQUIT SIGTERM (#44)
* [RES/SOCKET] HOTFIX | resolved the socket problome

* fix: change allow_methods to POST in default.conf

The allow_methods directive in the default.conf file was changed from GET to POST. This allows the server to accept POST requests in addition to GET requests.

* remove exe file

* [SERVER] feat: Remove debug print statements and update server socket backlog

The debug print statements in the `Client.cpp` and `ResponseHandle.cpp` files were commented out to remove unnecessary console output. Additionally, the server socket backlog in the `ServerSocket.cpp` file was updated to use the `SOMAXCONN` constant, which allows for a larger number of pending connections.

* [ERROR] refactor | Update logging level and error handling in code

The logging level in the code was updated to use the INFO level instead of the previous DEBUG level. Additionally, error handling was improved by adding error logging and throwing appropriate exceptions in certain scenarios.

* [SERVER] feat | Update signal handling in main.cpp

Improved signal handling in the `main.cpp` file by ignoring the `SIGINT`, `SIGKILL`, `SIGQUIT`, and `SIGTERM` signals. This ensures that the program does not terminate abruptly when these signals are received.